Translation API Key successfully read in from configuration properties file.
